requirements.txt
streamlit
supabase
google-generativeai
streamlit-supabase-auth
panda
cryptography